% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsrr_all_files.R
\name{nsrr_all_dataset_files}
\alias{nsrr_all_dataset_files}
\title{NSRR All Dataset files}
\usage{
nsrr_all_dataset_files(dataset, token = nsrr_token())
}
\arguments{
\item{dataset}{a dataset \code{"slug"}, one from
\code{\link{nsrr_datasets}}}

\item{token}{Token for NSRR resources.  Found at
\url{https://sleepdata.org/token}}
}
\value{
A \code{data.frame} of the data sets and their endpoints
}
\description{
NSRR All Dataset files
}
\examples{
\dontrun{
dataset = "shhs"
df = nsrr_all_dataset_files(dataset)
}
}
