% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsrr_website.R
\name{nsrr_api}
\alias{nsrr_api}
\title{NSRR API Functionality}
\usage{
nsrr_api(path, query = list(), token = NULL, ...)
}
\arguments{
\item{path}{path to endpoint for NSRR API}

\item{query}{query to pass to \code{\link{GET}}}

\item{token}{Authorization Token for NSRR, passed to
\code{\link{nsrr_token}}}

\item{...}{additional arguments to pass to \code{\link{GET}}}
}
\value{
A response object
}
\description{
NSRR API Functionality
}
