% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsrr_datasets.R
\name{nsrr_datasets}
\alias{nsrr_datasets}
\title{NSRR data sets}
\usage{
nsrr_datasets(token = nsrr_token())
}
\arguments{
\item{token}{Token for NSRR resources.  Found at
\url{https://sleepdata.org/token}}
}
\value{
A \code{data.frame} of the data sets and their endpoints
}
\description{
NSRR data sets
}
\examples{
df = nsrr_datasets()
testthat::expect_is(df, "data.frame")
}
