% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsrr_version.R
\name{nsrr_version}
\alias{nsrr_version}
\alias{nsrr_online_version}
\alias{nsrr_check_version}
\title{NSRR Gem version online}
\usage{
nsrr_version()

nsrr_online_version()

nsrr_check_version()
}
\value{
A character of the version
}
\description{
NSRR Gem version online
}
\examples{
nsrr_version()
nsrr_online_version()
nsrr_check_version()

}
