% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsrr_token.R
\name{nsrr_token}
\alias{nsrr_token}
\alias{nsrr_have_token}
\alias{nsrr_auth}
\alias{nsrr_authenticated}
\title{NSRR Token}
\usage{
nsrr_token(token = NULL)

nsrr_have_token(token = NULL)

nsrr_auth(token = NULL)

nsrr_authenticated(token = NULL)
}
\arguments{
\item{token}{Token for NSRR resources.  Found at
\url{https://sleepdata.org/token}}
}
\value{
A character vector or NULL
}
\description{
NSRR Token
}
\examples{
is.null(nsrr_token())
if (!is.null(nsrr_token())) {
   res = nsrr_auth()
   res$authenticated
}
bad_res = nsrr_auth("")
bad_res$authenticated
}
